Initialize the project's basic design
Added a main file to run the program and designed a class to securely
open files. Also, added some common files for typing and utilities.

The SecureOpen class does its best to prevent symlink race conditions
on all platforms. Also, on some platforms it is able to exclusively
open the file to prevent editing while it is open.